Update Vulkan bindings, drop interface enum in favor of union, events
Updated the Vulkan bindings used to the most recent commit.

Dropped the reified enum and associated hashmap in wl-interface in
favour of a reified tagged union, and flat array.

Created Event union type to accomodate the interface change, begin on
rewriting the event iterator to work with this new tagged union event.

Broke everything, but getting back to functionality, lol